pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-art-detector"
version = "0.1.0"
description = "Research-style AI art detector with reproducible data, training, evaluation, and deployment workflows."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Kevin" }]
dependencies = [
  "Pillow>=10.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
ml = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "seaborn>=0.13",
  "torch>=2.3",
  "torchvision>=0.18",
]
api = [
  "fastapi>=0.115",
  "httpx>=0.28",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.30",
]
demo = [
  "requests>=2.32",
  "streamlit>=1.38",
]
onnx = [
  "onnx>=1.16",
  "onnxruntime>=1.18",
  "onnxscript>=0.2",
]
dev = [
  "mypy>=1.11",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.6",
]

[project.scripts]
aiad = "ai_art_detector.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
